+++ python/__init__.py
@@ -1,3 +1,0 @@
-from rstar_python import PyRTree
-
-__all__ = ['PyRTree']